# docs/static/openapi.yaml
openapi: 3.0.0
info:
  contact:
    name: AltMount
    url: https://github.com/javi11/altmount/issues
  description: REST API for AltMount — Usenet WebDAV server with NZB queue, health
    monitoring, ARR integration, FUSE/rclone mounting, and Stremio addon.
  license:
    name: MIT
  termsOfService: http://altmount.kipsilabs.top
  title: AltMount API
  version: "1.0"
paths:
...
  /files/export-batch:
    post:
      description: Exports multiple mounted files as NZBs packed in a single ZIP archive.
      requestBody:
        content:
          application/json:
            schema:
              properties:
                paths:
                  items:
                    type: string
                  type: array
              type: object
        description: Virtual paths to export
        required: true
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          content:
            application/zip:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
        "500":
          description: Internal Server Error
          content:
            application/zip:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
      security:
        - BearerAuth: []
      summary: Batch export files as NZBs
      tags:
        - Files
  /files/export-nzb:
    get:
      description: Exports the metadata of a mounted file back to a downloadable NZB file.
      parameters:
        - description: Virtual path to the file
          in: query
          name: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          content:
            application/x-nzb:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
        "500":
          description: Internal Server Error
          content:
            application/x-nzb:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
      security:
        - BearerAuth: []
      summary: Export file metadata as NZB
      tags:
        - Files
  /files/info:
    get:
      description: Returns metadata for a mounted NZB file including segment info,
        encryption, and status.
      parameters:
        - description: Virtual path to the file
          in: query
          name: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/api.APIResponse"
                  - properties:
                      data:
                        $ref: "#/components/schemas/api.FileMetadataResponse"
                    type: object
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
      security:
        - BearerAuth: []
      summary: Get file metadata
      tags:
        - Files
...
  /import/file:
    post:
      description: Adds an NZB file at a given filesystem path to the import queue.
        Public endpoint (API key in query).
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/api.ManualImportRequest"
        description: File path to import
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/api.APIResponse"
                  - properties:
                      data:
                        $ref: "#/components/schemas/api.ManualImportResponse"
                    type: object
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
      security:
        - ApiKeyAuth: []
      summary: Manually import a file
      tags:
        - Import
...
servers:
  - url: //localhost:8080/api
components:
...
  securitySchemes:
    ApiKeyAuth:
      in: query
      name: apikey
      type: apiKey
  schemas:
    api.APIError:
      properties:
        code:
          type: string
        details:
          type: string
        message:
          type: string
      type: object
    api.APIMeta:
      properties:
        count:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        total:
          type: integer
      type: object
    api.APIResponse:
      properties:
        data: {}
        error:
          $ref: "#/components/schemas/api.APIError"
        meta:
          $ref: "#/components/schemas/api.APIMeta"
        success:
          type: boolean
      type: object
...
    api.FileMetadataResponse:
      properties:
        available_segments:
          type: integer
        created_at:
          type: string
        encryption:
          type: string
        file_size:
          type: integer
        modified_at:
          type: string
        nested_sources:
          items:
            $ref: "#/components/schemas/api.NestedSourceResponse"
          type: array
        password_protected:
          type: boolean
        segment_count:
          type: integer
        segments:
          items:
            $ref: "#/components/schemas/api.SegmentInfoResponse"
          type: array
        source_nzb_path:
          type: string
        status:
          type: string
      type: object
...
    api.ManualImportRequest:
      properties:
        file_path:
          type: string
        relative_path:
          type: string
      type: object
    api.ManualImportResponse:
      properties:
        message:
          type: string
        queue_id:
          type: integer
      type: object
...
    api.NestedSegmentResponse:
      properties:
        end_offset:
          type: integer
        message_id:
          type: string
        segment_size:
          type: integer
        start_offset:
          type: integer
      type: object
    api.NestedSourceResponse:
      properties:
        encrypted:
          type: boolean
        inner_length:
          type: integer
        inner_volume_size:
          type: integer
        segment_count:
          type: integer
        segments:
          items:
            $ref: "#/components/schemas/api.NestedSegmentResponse"
          type: array
        volume_index:
          type: integer
      type: object
...
    api.SegmentInfoResponse:
      properties:
        available:
          type: boolean
        end_offset:
          type: integer
        message_id:
          type: string
        segment_size:
          type: integer
        start_offset:
          type: integer
      type: object
...